chore: consolidate v1 root modules into utils
Move scoring, push, artifacts, decorators, loaders, and retries from the
v1 package root into verifiers/v1/utils, and inline the sample() helper
at its two call sites (SEED now lives next to Taskset.shuffle).
