+++ verifiers/v1/configs/cli/debug.py
@@ -52,7 +52,7 @@ class DebugConfig(BaseConfig):
     @model_validator(mode="before")
     @classmethod
     def resolve_taskset(cls, data):
-        from verifiers.v1.loaders import narrow_plugin_field, taskset_config_type
+        from verifiers.v1.utils.loaders import narrow_plugin_field, taskset_config_type
 
         narrow_plugin_field(data, "taskset", taskset_config_type)
         return data
