+++ verifiers/v1/mcp/toolset.py
@@ -4,10 +4,10 @@ from typing import TYPE_CHECKING
 
 from pydantic_config import BaseConfig
 
-from verifiers.v1.decorators import discover_decorated
 from verifiers.v1.mcp.server import ConfigT, ServerBase
 from verifiers.v1.runtimes import RuntimeConfig, SubprocessConfig
 from verifiers.v1.state import StateT
+from verifiers.v1.utils.decorators import discover_decorated
 
 if TYPE_CHECKING:
     from mcp.server.fastmcp import FastMCP
