+++ verifiers/v1/state.py
@@ -1,8 +1,4 @@
-"""Mutable state shared within one rollout.
-
-Tool servers synchronize it through the interception state channel. It is excluded
-from serialized traces.
-"""
+"""Mutable rollout-level state shared across tool server + host."""
 
 from pydantic import BaseModel, ConfigDict, Field
 from typing_extensions import TypeVar
@@ -12,6 +8,7 @@ from verifiers.v1.utils.generic import concrete_type
 
 class State(BaseModel):
     model_config = ConfigDict(ser_json_inf_nan="constants")
+
     artifacts: dict[str, bytes] = Field(default_factory=dict)
