+++ verifiers/v1/utils/artifacts.py
