+++ verifiers/v1/utils/platform.py
